requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1; python_version < '3.9'
dj-rest-auth==6.0.0
django-cors-headers==4.4.0
django==4.2.14
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
pyjwt==2.8.0
pytz==2024.1
pyyaml==6.0.1
sqlparse==0.5.1
typing-extensions==4.12.2; python_version < '3.11'
uritemplate==4.1.1
